fixed casting double to long bug

+++ app/src/main/java/com/cmput301f21t34/habittrak/DatabaseManager.java
@@ -386,8 +386,8 @@ public class DatabaseManager {
 
         if(hashMap != null) {
             Location loc = new Location((String) hashMap.get("provider"));
-            loc.setLatitude((double) (long) hashMap.get("latitude"));
-            loc.setLongitude((double) (long) hashMap.get("longitude"));
+            loc.setLatitude((double) hashMap.get("latitude"));
+            loc.setLongitude((double)  hashMap.get("longitude"));
 
             return loc;
         }
